requirements.txt
langchain
langchain_experimental
langchain_core
langchain_google_genai
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
langchain-google-genai
langchain[docarray]
transformers
colorama
beautifulsoup4
bs4
lxml
PyPDF2